+++ resolwe_bio/kb/management/commands/insert_features.py
@@ -8,6 +8,7 @@ Insert Knowledge Base Features
 import csv
 import json
 import logging
+import re
 
 from django.core.exceptions import ValidationError
 from django.core.management.base import BaseCommand
@@ -85,7 +86,11 @@ class Command(BaseCommand):
         parser.add_argument(
             "file_name",
             type=str,
-            help="Tab-separated file with features (supports tab, gz or zip)",
+            help=(
+                "Tab-separated file with features (supports tab, gz or zip)."
+                "When file_name starts with s3:// it is downloaded from S3 "
+                "via boto3 library. See also --region."
+            ),
         )
 
     def handle(self, *args, **options):
@@ -95,6 +100,21 @@ class Command(BaseCommand):
 
         type_choices = list(zip(*Feature.TYPE_CHOICES))[0]
         subtype_choices = list(zip(*Feature.SUBTYPE_CHOICES))[0]
+
+        # Download file if it is located on S3.
+        match = re.search(r"^s3://([a-z0-9-.]{3,63})/(.*)$", options["file_name"])
+        if match:
+            try:
+                import boto3
+            except ImportError:
+                logger.exception(
+                    "Package 'boto3' must be installed to download data from S3 bucket."
+                )
+                raise
+
+            bucket, key = match.groups()
+            boto3.resource("s3").Bucket(bucket).download_file(key, key)
+            options["file_name"] = key
 
         for tab_file_name, tab_file in decompress(options["file_name"]):
             logger.info(__('Importing features from "{}"...', tab_file_name))